male upravy - odstranena cestina z controlleru

+++ app/controllers/admin_controller.rb
@@ -221,10 +221,14 @@ class AdminController < ApplicationController
 	end
 
 	def delete_events_and_race
-		Event.delete_all
-		Race.delete_all
+    if Time.now < $dc.start_time
+		  Event.delete_all
+		  Race.delete_all
 
-		flash[:notice] = "Všechny Eventy a Race tabulka byly odstraněny!"
+		  flash[:notice] = "All events and race table was removed!"
+    else
+		  flash[:alert] = "Events can be deleted only before race starts!"
+    end
 		redirect_to :action => 'results_setting'
 	end
